173: apply review feedback
