+++ lib/features/campaigns/models/marker_item_model.dart
