+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   flutter_native_splash: ^2.4.2
   webview_flutter: ^4.10.0
   maplibre_gl: ^0.20.0
+  maplibre_gl_platform_interface: ^0.20.0
 #--------- swagger_dart_code_generator --------------
   chopper: ^8.0.0
   json_annotation: ^4.8.0
@@ -43,10 +44,9 @@ dependencies:
   http: ^1.2.2
   turf: ^0.0.10
   camerawesome: ^2.1.0
-  http_parser: ^4.0.2
   intl: ^0.19.0
-  maplibre_gl_platform_interface: ^0.20.0
   image: ^4.3.0
+  http_parser: ^4.0.2
 
 dev_dependencies:
